requirements.txt
pydub==0.25.1
python-dotenv==1.0.1
python-ESL==1.4.18
